+++ jwst/lib/tests/engdb_mock.py
@@ -7,7 +7,7 @@ import os
 import re
 import requests_mock
 
-from jwst.lib import engdb_tools
+from jwst.lib import engdb_direct, engdb_tools
 
 __all__ = [
     'ENGDB_PATH',
@@ -63,14 +63,14 @@ class EngDB_Mocker(requests_mock.Mocker):
 
         # Setup from meta query
         meta_query = re.compile(''.join([
-            engdb_tools.ENGDB_METADATA,
+            engdb_direct.ENGDB_METADATA,
             '.*'
         ]))
         self.get(meta_query, json=self.response_meta)
 
         # Setup to return a general data query
         data_query = re.compile(''.join([
-            engdb_tools.ENGDB_DATA,
+            engdb_direct.ENGDB_DATA,
             r'.+\?sTime=.+\&eTime=.+'
         ]))
         self.get(data_query, json=self.response_data)
@@ -129,6 +129,23 @@ class EngDB_Mocker(requests_mock.Mocker):
         context.status_code = 200
         return data
 
+    def __enter__(self):
+        """Setup environment for the context
+
+        Remove MAST_API_TOKEN to ensure the EngdbMast service is not used.
+        """
+        self._environ = os.environ.copy()
+        try:
+            del os.environ['MAST_API_TOKEN']
+        except KeyError:
+            pass
+        super().__enter__()
+
+    def __exit__(self, type, value, traceback):
+        """Restore the environment"""
+        super().__exit__(type, value, traceback)
+        os.environ = self._environ
+
 
 class EngDB_Local():
     """
@@ -173,13 +190,13 @@ class EngDB_Local():
         data = db_data['Data']
         start_idx = 0
         try:
-            while engdb_tools.extract_db_time(data[start_idx]['ObsTime']) < stime_mil:
+            while engdb_direct.extract_db_time(data[start_idx]['ObsTime']) < stime_mil:
                 start_idx += 1
         except IndexError:
             pass
         end_idx = start_idx
         try:
-            while engdb_tools.extract_db_time(data[end_idx]['ObsTime']) <= etime_mil:
+            while engdb_direct.extract_db_time(data[end_idx]['ObsTime']) <= etime_mil:
                 end_idx += 1
         except IndexError:
             end_idx -= 1
@@ -300,7 +317,7 @@ def cache_engdb(
         json.dump(meta, fp)
 
     for mnemonic in mnemonics:
-        records = edb.get_records(mnemonic, starttime, endtime)
+        records = edb._get_records(mnemonic, starttime, endtime)
 
         # Remove the request times. These are filled back in
         # during retrieval.